PBI #32580 Remove self-update feature from name search input At the beginning we have implemented keypress aware input box for name-search that will send a data update request when the user has paused typing for a short while, but will stay silent if the user keeps on typing into the input.
But this fails for a few cases:
1. We a change() event is triggered through javascript, like in web behavior test.
2. When the user types chinese chars using input helper, such that the typed content is buffered in other softwares.
3. When the user types chinese chars using input helper, and use a combination of keypress and mouse click.

It's hard to solve all of them at the same time with our initial approach.
So instead of auto-refreshing on delayed keypress, we now require the user to actively
invoke change() by pressing enter or defocusing the input box, or click the search button
beside the search box.

In addition, we adjust the sql statement a little bit to allow searching for empty space inside the name text.