Fix build.
Ignore this test because it doesn't test anything.

But it is worth mentioning that it fails because I cannot create a valid
  TeleoptiPrincipal.CurrentPrincipal
in in the test fake. It is a static method that returns Thread.CurrentPrincipal().
I don't know how to inject such that TeleoptiPrincipal.CurrentPrincipal() doesn't
return null in WebTest.

I will come back to this when I find the way to do it.