Fix outbound scenario (again++).
This time the timeout does have a reasonable cause. The workload day template is not determined when setting up campaign workload with configurables.
But they have to be set up when the campaign is editted, which will use significant amount of time.

It's enough to reduce the campaign duration to suppress this issue without altering the original test purpose.