Removing (probably) not needed VersionStateRollbackInterceptor.
Had to remove one test when saving entities and transaction is rollbacked.
This no longer resets Id and Version. AFAIK, this is only suppose
to work in scheduler (when conflicts) and there we have test/impl code for this in
our persisters instead.

A bit dangerous so it's done behind toggle
ResourcePlanner_RemoveOldInterceptorProbablyNotNeeded_00000
(immediately is set to true)

Talked with Anders - to avoid extra work, no explicit PBI is done
for this push.