Merged PR 1383: story74957 - frontendtests
note that 'should submit form', is not tested..
any ideas how that could be possible?

I guess it could be created with some sort of jasmine.createSpyObj, but in the end it feels nothing really gets tested, if i do. So i'm not really sure if it should be tested :).

Related work items: #74957