Merged PR 2002: ##82152 Make shift lengths in be read as hours and minutes (7 h 30 min) instead of as timestamp (8:30)
In the MyTime Schedule week and month view we currently display the shift lengths as 8:30. This is supposed to mean "The shift is 8 hours and 30 minutes", but is read as "Eight-thirty" or "Half past eight o clock" by screen readers. It's not very obvious to sighted users either.

This PR changes the shift length to use the format "X h Y min" (e.g. "7 h 30 min"). This works well with screen readers. The hour and minutes abbreviations are localized.

I have solved this by adding a JS function to the model files for the week view and the month view. This could have been done further backend (in _Timehelper.cs_, but I'm not very comfortable with C# and didn't know how to access the usertexts from there.

Related work items: #82152